requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pycocotools>=2.0.6
pydicom>=2.3.0
Pillow>=9.0.0
numpy>=1.24.0
pandas>=1.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
scipy>=1.10.0
tqdm>=4.65.0
albumentations>=1.3.0
opencv-python>=4.7.0
kaggle>=1.5.0
timm>=0.9.0
wandb>=0.15.0
